aria-backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
groq==0.11.0
python-dotenv==1.0.1
websockets==12.0
requests==2.32.0
